商品管理1

+++ src/main/java/com/team07/online_shopping_mall/web/controller/ProductController.java
@@ -1,5 +1,10 @@
 package com.team07.online_shopping_mall.web.controller;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.team07.online_shopping_mall.mapper.ProductMapper;
+import com.team07.online_shopping_mall.model.domain.User;
+import com.team07.online_shopping_mall.service.UserService;
+import org.apache.ibatis.annotations.Param;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.stereotype.Controller;
 import org.slf4j.Logger;
@@ -9,6 +14,9 @@ import org.springframework.web.bind.annotation.*;
 import com.team07.online_shopping_mall.common.JsonResponse;
 import com.team07.online_shopping_mall.service.ProductService;
 import com.team07.online_shopping_mall.model.domain.Product;
+
+import java.util.ArrayList;
+import java.util.List;
 
 
 /**
@@ -29,6 +37,11 @@ public class ProductController {
     @Autowired
     private ProductService productService;
 
+    @Autowired
+    private ProductMapper productMapper;
+    @Autowired
+    private UserService userService;
+
     /**
     * 描述：根据Id 查询
     *
@@ -46,8 +59,10 @@ public class ProductController {
     */
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     @ResponseBody
-    public JsonResponse deleteById(@PathVariable("id") Long id) throws Exception {
-        productService.removeById(id);
+    public JsonResponse deleteById(@RequestParam("user") User user, @RequestParam("id") Long id) throws Exception {
+        if(productService.identifyUser(user,productService.getById(id).getShopId())){
+            productService.removeById(id);
+        }
         return JsonResponse.success(null);
     }
 
@@ -58,10 +73,34 @@ public class ProductController {
     */
     @RequestMapping(value = "", method = RequestMethod.PUT)
     @ResponseBody
-    public JsonResponse updateProduct(Product  product) throws Exception {
-        productService.updateById(product);
+    public JsonResponse updateProduct(@RequestParam("user") User user, @RequestParam("product") Product  product) throws Exception {
+//        User user1 =userService.getById(3);
+//        Product product1=productService.getById(2);
+//        product1.setDetail("非常好喝1");
+        //product1.setId();
+        if(productService.identifyUser(user,product.getShopId())){
+            productService.updateById(product);
+            //System.out.println(1);
+        }
         return JsonResponse.success(null);
     }
+
+
+
+
+    /**
+     * 描述：根据product_name 查询所有
+     *
+     */
+    @RequestMapping(value = "/name")
+    @ResponseBody
+    public JsonResponse getByName(@RequestParam String name)throws Exception {
+        List<Product> lists=productMapper.getByName(name);
+        return JsonResponse.success(lists);
+    }
+
+
+
 
 
     /**
@@ -71,6 +110,8 @@ public class ProductController {
     @RequestMapping(value = "", method = RequestMethod.POST)
     @ResponseBody
     public JsonResponse create(Product  product) throws Exception {
+//        Product product2=new Product();
+//        product2=productService.getById(2L);
         productService.save(product);
         return JsonResponse.success(null);
     }
